多语言下拉框挪到其他设置下
